edit: deleted uint64_t usages because architecture is now 32 bits, toolchain, new: kmain uart init

+++ includes/standard.h
@@ -7,9 +7,9 @@
 
 typedef unsigned int uint_t;
 
-uint64_t strlen(const uint8_t *s);
+uint32_t strlen(const uint8_t *s);
 int strcmp(const uint8_t *s1, const uint8_t *s2);
-int strncmp(const uint8_t *s1, const uint8_t *s2, uint64_t len);
-void *memset(void *s, int c, uint64_t n);
+int strncmp(const uint8_t *s1, const uint8_t *s2, uint32_t len);
+void *memset(void *s, int c, uint32_t n);
 
 #endif
